+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 - Upgrade `phpunit/phpunit` to `^11.5` ([#414](https://github.com/opensearch-project/opensearch-php/pull/414))
 ### Deprecated
+- Passing a `CredentialProvider` instance to `SigningClientFactory` is deprecated; pass a `callable` instead ([#404](https://github.com/opensearch-project/opensearch-php/issues/404))
 ### Removed
 - Drop CI integration testing against EOL OpenSearch 1.x and legacy Elasticsearch OSS 7.10.0
 ### Fixed